refactor(validation): unify validated-field reads behind one resolver #1016
Addresses PR #1022 review: the magic-property feature was implemented as
a parallel subsystem instead of another spelling of an existing
"validated field read".

- Extract ValidatedFieldReadResolver + ValidatedFieldRead value object as
  the single place that answers "is this expression a validated field
  read, and which rule governs it?" for every syntax: keyed accessor
  methods, ValidatedInput accessors, magic property fetches, and tracked
  inline-validate variables. addTaints/removeTaints each ask once and read
  one facet (source vs escape); the per-syntax branches are gone.
- ValidationTaintHandler shrinks to pure taint-graph mechanism (source
  re-emission, per-node dedupe, cleanup), 785 -> 206 lines.
- Drop the assignment-LHS tracking machinery (BeforeExpressionAnalysis
  hook + marker set). A declared property opts out of narrowing, and an
  undeclared magic property cannot be written (Request has no __set, so
  the write is already UndefinedPropertyAssignment); the guard defended an
  unreachable state, and a missed write is a false negative that aligns
  with "silence over false positives". Verified: the full taint suite is
  unchanged with it removed.

Re-sourcing stays required (not replaceable by a __get stub source):
provider-supplied property types bypass the magic method, so a
@psalm-taint-source on __get does not fire for $req->email (confirmed
empirically). Dedupe stays required: the read pass and argument-binding
pass dispatch the same node twice.
